Ensure the .deps file is included in debian source
This (empty) file is needed to build the dep. Without it, you get this
error:

    dh clean --with apache2
    dh: warning: Compatibility levels before 10 are deprecated (level 9 in use)
       debian/rules override_dh_auto_clean
    make[1]: Entering directory '/home/amanda/third-party-software/apache2-mod-log-ipmask'
    make clean
    make[2]: Entering directory '/home/amanda/third-party-software/apache2-mod-log-ipmask'
    /usr/share/apache2/build/rules.mk:236: .deps: No such file or directory
    make[2]: *** No rule to make target '.deps'.  Stop.
    make[2]: Leaving directory '/home/amanda/third-party-software/apache2-mod-log-ipmask'
    make[1]: *** [debian/rules:7: override_dh_auto_clean] Error 2
    make[1]: Leaving directory '/home/amanda/third-party-software/apache2-mod-log-ipmask'
    make: *** [debian/rules:4: clean] Error 2
    dpkg-buildpackage: error: fakeroot debian/rules clean subprocess returned exit status 2

The deb file can be built with `dpkg-buildpackage` without this change.
However if you want to build a deb using cowbuilder for a different
distribution, you need to create a source package and `.dsc` file. The
`.deps` file will be excluded by default by dpkg, and the deb file will
fail to build.

The fix here is to overwrite the dpkg-source “don't include these files”
argument to remove the `.deps`.